requirements.txt
numpy
pandas
matplotlib
scipy
setuptools
scikit-learn
python-dotenv
efficient-apriori
matplotlib
mpl_toolkits
